requirements.txt
streamlit>=1.25.0
openai>=0.27.0
cryptography>=40.0.0
python-dotenv>=1.0.0
